requirements.txt
Django==3.0.3
djangorestframework==3.11.0
drf-access-policy==0.5.1
django-rest-knox==4.1.0
